requirements.txt
flask
opencv-python-headless
numpy
torch
torchvision
transformers
pillow
requests
yt-dlp
werkzeug
